update math functions

+++ Core/src/Physics/include/Physics/PhysicsTypes.hpp
@@ -52,9 +52,9 @@ namespace Vakol
         Math::Vec3 worldNormal = Math::Vec3(0.0f, 0.0f, 0.0f);
         Math::Vec3 worldPoint = Math::Vec3(0.0f, 0.0f, 0.0f);
         Math::Vec3 localPoint = Math::Vec3(0.0f, 0.0f, 0.0f);
-        double penetrationDepth = 0.0;
+        float penetrationDepth = 0.0;
         bool isColliding = false;
-        double lambda = 0.0;
+        float lambda = 0.0;
 
         RigidBody* parentBody = nullptr;
     };
@@ -66,7 +66,7 @@ namespace Vakol
     struct RigidBody
     {
         BodyType type = BodyType::Dynamic;
-        double mass = 1.0;
+        float mass = 1.0;
         // double bounciness = 0.3;
         Math::Vec3 bounciness = Math::Vec3(0.8f, 0.3f, 0.8f);
         bool hasGravity = true;
